ui/requirements.txt
streamlit==1.42.0
requests==2.32.3
Pillow==11.1.0
boto3